+++ src/deluge/gui/ui/menus.cpp
@@ -105,6 +105,7 @@
 #include "gui/menu_item/midi/mpe_to_mono.h"
 #include "gui/menu_item/midi/pgm.h"
 #include "gui/menu_item/midi/program.h"
+#include "gui/menu_item/midi/kit/output_channel.h"
 #include "gui/menu_item/midi/sound/channel.h"
 #include "gui/menu_item/midi/sound/note_for_drum.h"
 #include "gui/menu_item/midi/sub.h"
@@ -621,6 +622,10 @@ HorizontalMenu soundDistortionMenu{
 midi::sound::OutputMidiChannel outputMidiChannelMenu{STRING_FOR_CHANNEL, STRING_FOR_CHANNEL};
 midi::sound::OutputMidiNoteForDrum outputMidiNoteForDrumMenu{STRING_FOR_NOTE, STRING_FOR_NOTE};
 Submenu outputMidiSubmenu{STRING_FOR_MIDI, {&outputMidiChannelMenu, &outputMidiNoteForDrumMenu}};
+
+// Kit-level MIDI output: sets channel + sequential note numbers for all SoundDrums at once
+midi::kit::OutputMidiChannel kitOutputMidiChannelMenu{STRING_FOR_CHANNEL, STRING_FOR_CHANNEL};
+Submenu kitMidiOutputSubmenu{STRING_FOR_MIDI, {&kitOutputMidiChannelMenu}};
 
 // MIDIInstrument menu ----------------------------------------------------------------------
 midi::device_definition::Linked midiDeviceLinkedMenu{STRING_FOR_MIDI_DEVICE_DEFINITION_LINKED,
@@ -1675,6 +1680,7 @@ menu_item::Submenu soundEditorRootMenuKitGlobalFX{
     {
         &kitGlobalFXActionsMenu,
         &kitClipMasterMenu,
+        &kitMidiOutputSubmenu,
         &kitArpMenu,
         &audioCompMenu,
         &globalFiltersMenu,